// src/client/schemas.ts
export enum WebcastSchemaVersion {
  SCHEMA_VERSION_UNSPECIFIED = 0,
  SCHEMA_VERSION_V1 = 1,
  SCHEMA_VERSION_V2 = 2,
  SCHEMA_VERSION_V3 = 3,
  UNRECOGNIZED = -1,
}

export enum WebcastPlatform {
  WEBCAST_PLATFORM_UNSPECIFIED = 0,
  WEBCAST_PLATFORM_WEB = 1,
  WEBCAST_PLATFORM_MOBILE = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface WebSocketOptions {
  uniqueId: string;
  jwtKey?: string | undefined;
  apiKey?: string | undefined;
  xCookieHeader?: string | undefined;
  xOauthToken?:
    | string
    | undefined;
  /** WebSocket feature flags */
  features:
    | WebSocketFeatureFlags
    | undefined;
  /** WebSocket gateway flags */
  flags: WebSocketGatewayFlags | undefined;
}

export interface WebSocketGatewayFlags {
  /** Whether the gateway will bundle events or emit individually */
  bundleEvents?:
    | boolean
    | undefined;
  /** The output schema version */
  schemaVersion?:
    | WebcastSchemaVersion
    | undefined;
  /** Whether the gateway will output raw binary or JSON */
  rawMessages?:
    | boolean
    | undefined;
  /** Whether to include raw bytes as base64 in the JSON when raw_messages = false */
  includeRawBytes?: boolean | undefined;
}

export interface WebSocketFeatureFlags {
  /** Whether to normalize the Unique ID value */
  normalizeUniqueId?:
    | boolean
    | undefined;
  /** Whether to use the Enterprise API when connecting */
  useEnterpriseApi?:
    | boolean
    | undefined;
  /** Platform to use */
  webcastPlatform: WebcastPlatform;
  /**
   * Whether the gateway should also attach the client to the peer
   * transcription stream for this creator.
   */
  enableTranscription?: boolean | undefined;
}